pyproject.toml
[tool.poetry]
name = "transaction-analysis"
version = "0.1.0"
description = ""
authors = ["Nikolay_Lyalkin <serega94nn@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.11.0"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pandas = "^2.2.2"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.venv'

[tool.isort]
line_length = 119
